+++ packages/sdk-governance/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.3.0 2025-05-27
+
 ### Added
 
 - **ChildBounties**